refactor(renderer): 根级 monaco/markdown/editor-font 收敛到 lib/
renderer/src 根目录的三个第三方渲染/编辑器集成初始化文件（monaco-setup /
markdown / editor-font）收进 lib/，根目录只剩入口与 api/errors。纯文件位置调整
+ import 路径改写，无行为变更。
