+++ apps/desktop/src/renderer/src/lib/monaco-setup.ts
